feat: allow inclusion of 'examples' keyword
In addition to the existing support for a route
'example', this commit allows for the specification
of an 'examples' keyword for multiple examples.
As JSONSchema requests examples in an Array, and
OpenAPI spec an object, we extract the 'name'
property from each array entry and use this as
the key for the OpenAPI spec object.

Refs: #26